Added profiles for with and without GPU.
This commit adds multiple profiles to the docker-compose file. The ''
(empty) profile is the default profile (which is ran when using
`docker-compose up`). the "gpu" profile runs the same services, except
the experimentation service is ran with GPU enabled. Finally, MLFlow has
a separate profile to be ran as standalone service (when running the
services locally).