Changed dataset handling and implemented downloading artifacts.
Changed the way datasets are handled from copying the datasets to a temp
folder to using the dataset as is in the original location, but
providing it in a read-only way (so that a dataset cannot be changed by
one task while it is being used by another task (in this case a
temporary directory is provided). Additionally, a download production
model method has been added to the connector as a way to download the
latest model.